requirements.txt
# Django & DRF
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter>=23.5
django-cors-headers>=4.3
django-environ>=0.11

# Database
psycopg2-binary>=2.9

# Auth & Security
PyJWT>=2.8
cryptography>=41.0

# Redis & Celery
redis>=5.0
django-redis>=5.4
celery[redis]>=5.3
django-celery-beat>=2.5
django-celery-results>=2.6

# OTP & Utils
openpyxl>=3.1
pyotp>=2.9
phonenumbers>=8.13
requests>=2.31
twilio>=8.0

# Media & Storage (S3/MinIO)
Pillow>=10.0
boto3>=1.34
django-storages[s3]>=1.14

# Channels (WebSocket)
channels>=4.0
channels-redis>=4.2
daphne>=4.0

# Production
gunicorn>=21.0
whitenoise>=6.6
